+++ models/posts.js
@@ -51,9 +51,10 @@ const PostsSchema           = new mongoose.Schema({
 
 
 // ==============================
-//  STATIC METHODS
+//  METHODS
 // ==============================
-// (pre-hook) version counter
+// action hooks
+// version counter (pre-hook)
 PostsSchema.pre('findOneAndUpdate', function () {
     this.findOneAndUpdate({}, { $inc: { _revised: 1 }});
 });